requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.1
alembic==1.13.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.2
contourpy==1.1.1
cycler==0.12.1
docker==7.0.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==21.2.0
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.1.2
lightning-utilities==0.10.1
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.5
mlflow==2.11.1
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
networkx==3.1
numpy==1.24.4
opencv-python-headless==4.9.0.80
optuna==3.5.0
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pretrainedmodels==0.7.4
protobuf==4.25.3
pyarrow==15.0.0
pydicom==2.4.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.3.0
tifffile==2023.7.10
timm==0.9.2
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.1+cu118
torchmetrics==1.2.1
torchvision==0.16.1+cu118
tqdm==4.66.2
triton==2.1.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
